update kubeflow-dashboard-links to final schema
