+++ interfaces/kubeflow_dashboard_links/v0/README.md
@@ -1,19 +1,19 @@
-# `kubeflow-dashboard-sidebar/v0`
+# `kubeflow-dashboard-links/v0`
 
 ## Usage
 
 This relation interface describes the expected behavior of any charm claiming to be able to provide 
-or consume the `kubeflow-dashboard-sidebar`.  `kubeflow-dashboard-sidebar` is an interface for defining links in the format of Kubeflow Dashboard's sidebar. 
+or consume the `kubeflow-dashboard-links` interface.  `kubeflow-dashboard-links` is an interface for defining links in the format needed by Kubeflow Dashboard. 
 
 ## Behavior
 
 ### Provider
 
-- Is expected to create a sidebar link in the dashboard UI for each link entry sent by the `requirer` over the relation
+- Is expected to create a link in the dashboard UI for each link entry sent by the `requirer` over the relation
 
 ### Requirer
 
-- Is expected to send zero or more sidebar items to the `provider` in the below-defined format. 
+- Is expected to send zero or more links to the `provider` in the below-defined format. 
 
 ## Relation Data
 
@@ -30,14 +30,18 @@ The requirer specifies zero or more sidebar items in the required format.
   {
     "text": "Some link text",
     "link": "/some-relative-link",
+    "location": "menu",
     "type": "item",
-    "icon": "assessment"
+    "icon": "assessment",
+    "desc": "link description"
   },
   {
     "text": "Some other link text",
     "link": "/some-other-relative-link",
+    "location": "quick",
     "type": "item",
-    "icon": "book"
+    "icon": "book",
+    "desc": "link description"
   }
 
 ]
