+++ interfaces/kubeflow_dashboard_links/v0/charms.yaml
